requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0
joblib>=1.2.0
PyYAML>=6.0
streamlit>=1.25.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0